requirements.txt
librosa
dtw
matplotlib
pyacoustid
python-dotenv
mongoengine
cursor
pathlib
